Added style for calendar
Previously the Mid-se break and SWOT VAC rows in Unit Calendar were styled using a cell ID. But this was not enough since this style would disappear if lecturers deleted or rearranged rows. Now we use a js to style any row with the content "mid-semester break" or "swot vac" in it.