Fixed Accessibility Issue
Anddy reported an issue where participant names and other links in the participants table were displaying in white and thus not very visible. The new css amendments make it so that participant table is not affected by our styling and so will display in blue text (default for linked text).